I made a blank of the Home page and registered a route for it.  Test for   zenbit.tech
I made a blank of the Home page and registered a route for it. To do this, in the Home.js file, in the Home function, I returned some markup for the Home page blank and exported this Home function. Then, in the index.js file, I re-exported the Home page. Then, in the App.jsx file, I imported Routes, Route and lazy. I also imported the Home component with help the lazy function In the App function, I returned <Routes> in which I specified the Route where I specified the <Home />     page.  Test for   zenbit.tech