Added navigation buttons My Logo and Catalog and slightly implemented the logic of their operation.   Test for   zenbit.tech
Added navigation buttons My Logo and Catalog and slightly implemented the logic of their operation. To do this, I wrote a blank for the Catalog component in the Catalog.jsx file. Then, in the index.js file (which is located in the catalog folder), I re-exported the Catalog component. Then, in the Navigation.jsx file, I imported NavLink and, using it in the Navigation function, made transitions along routes when clicked to the My Logo or Catalog buttons And also re-exported the Navigation component in the index.js file (which is located in the navigation folder) Then, in the Layout.jsx file, I imported Suspense and Outlet to use them to render different page bodies in the Layout component. I also rendered the component
 Navigation in it. And also re-exported the Layout component in the index.js file (which is located in the layout folder). In the App.jsx file I commented out the Home component for now (I will uncomment it and render it in the next commits) I also imported the Layout component, and using lazy I imported the Catalog component and assigned routes to them in Route   .   Test for   zenbit.tech