I uncommented the import of the Home page in the App.jsx file and specified the path for it. I also filled the Home page with content and applied styles to it.  Test for   zenbit.tech
I uncommented the import of the Home page in the App.jsx file and specified the path for it. I also filled the Home page with content and applied styles to it.  To do this, I wrote styles for the content of the Home page in the Home.styled.js file. Please note that the background for this page does not load (something needs to be fixed). And then imported these styles into the Home.jsx file. Also adjusted the styles in the Navigation.styled.js file. Test for   zenbit.tech