Added Sign Out button to header.  Test for zenbit.tech
Added Sign Out button to header.  To do this, I wrote markup for the Sign Out button in the UserMenu.jsx file and imported into this file the styles that I wrote in the UserMenu.styled.js file. I also re-exported the UserMenu.jsx file using the index.js file, which is located in the userMenu folder. Then I adjusted the styles in file Navigation.styled.js Then called the Navigation and UserMenu components in the AppBar.jsx file. Also styled the AppBar component in the AppBar.styled.js file. And re-exported the AppBar.jsx file using the index.js file, which is located in the appBar folder. Also redid the Layout file. jsx by calling the AppBar component in it. Test for zenbit.tech