Added Log In and Sign Up buttons to the header.  Test for zenbit.tech
Added Log In and Sign Up buttons to the header.  To do this, I wrote a fake selector in the selectors.js file. It is not working, but in the future I will refine it and use it in this logic. Then I wrote markup in the AuthMenu.jsx file for the Log In and Sign Up buttons, having previously imported styles here from the AuthMenu.styled.js file. Then I re-exported the AuthMenu.jsx file using the index.js file, which is located in the authMenu folder. Then I corrected pathes for import of the Navigation and UserMenu components in the AppBar.jsx file  (they just could have been written a little shorter). Then I imported the AuthMenu component and also imported and immediately commented out useSelector and selectIsLoggedIn. I also commented out the isLoggedIn variable. I prepared the logic (but it’s not working yet) for the ternary in line 14   Test for zenbit.tech