I created a user registration page.  Test for zenbit.tech
I created a user registration page.  To do this, I imported useState and styles that were written in the RegisterForm.styled.js file into the RegisterForm.jsx file. I also imported the rectangle photo. Then, in the RegisterForm function, I set the initial values of the form fields using useState . Then, using switch, I wrote the value from the corresponding form field in the handleChange variable. Then I returned some markup. At the same time, pay attention, in order to display an image on the page, you need to display the image imported into this file in src, so that it turns out like this <img src={rectangle} alt="town" /> . In my case it turned out <StyledRegisterFormImg src={rectangle} alt="town" /> because I styled it using styled-components . Then I re-exported the RegisterForm.jsx file using the index.js file located in the registerForm folder. Then in the Navigation.jsx file I imported useSelector and the selectIsLoggedIn selector itself and then in the isLoggedIn variable I received the values ​​from this selector (this logic does not work yet. I will finalize it in the next commits). Also in this file I set the Catalog button to be displayed manually using the usual false . Then in the AuthMenu.jsx file I simply swapped the paths so that the paths corresponded to the buttons for navigating to those paths. Then in the App.jsx file I imported the RegisterForm component and Route assigned the path to it.  Test for zenbit.tech