I made a form for entering login and password.   Test for zenbit.tech
I made a form for entering login and password.  To do this, in the Login.jsx file I completely used the logic of the RegisterForm.jsx page,  I only removed the name field because it is not needed here. I also added styles in the Login.styled.js file and re-exported the Login.jsx file using the index.js file, which is located in the login folder. Then I imported the Login page into the App.jsx file using the lazy function and assigned a path to it in Route . Please note that in this file I removed the / from the path for the RegisterForm page because it is not needed here.  Test for zenbit.tech